bot/requirements.txt
beautifulsoup4==4.12.3
Pillow==10.3.0
python-dotenv==1.0.1
requests==2.32.3
selenium==4.21.0
